requirements.txt
Flask==3.1.0
protobuf==5.29.2
python-dotenv==1.0.1
